Organize files
- Play shoot pascal;
