+++ src/Objects/Platform/Power.js
@@ -1,4 +1,4 @@
-export default class PascalPower extends Phaser.Physics.Arcade.Sprite {
+export default class Power extends Phaser.Physics.Arcade.Sprite {
 	constructor(scene, x, y) {
 		super(scene, x, y, "PascalZim");
 		scene.add.existing(this);
