+++ src/Objects/Platform/Shoot.js
@@ -1,0 +1,5 @@
+export default class Shoot extends Phaser.Physics.Arcade.Sprite {
+	constructor(scene, x, y) {
+		super(scene, x, y, "PascalZim");
+	}
+}
