+++ src/Scenes/Options.js
@@ -3,8 +3,8 @@ import GlobalConfigs from "../Config/Configs";
 import AnimateTitle from "../Components/AnimatedTitle";
 import OptionsButton from "../Components/OptionsButton";
 
-import Player from "../Objects/Player";
-import Tiles from "../Objects/Tiles";
+import Player from "../Objects/Platform/Player";
+import Tiles from "../Objects/Platform/Tiles";
 import DirectionSign from "../Objects/DirectionSign";
 
 export default class Options extends Phaser.Scene {
